Move test data and mocks to template class
