+++ budgeteer-web-interface/src/test/java/org/wickedsource/budgeteer/web/pages/budgets/details/BudgetDetailsPageTest.java
@@ -14,29 +14,12 @@ import static org.mockito.Mockito.when;
 
 public class BudgetDetailsPageTest extends AbstractWebTestTemplate {
 
-    @Autowired
-    private BudgetService service;
-
     @Test
     void render() {
         WicketTester tester = getTester();
-        when(service.loadBudgetDetailData(1L)).thenReturn(createBudget());
-        when(service.loadBudgetBaseData(1L)).thenReturn(new BudgetBaseData(1L, "Budget 1"));
         tester.startPage(BudgetDetailsPage.class, BudgetDetailsPage.createParameters(1L));
         tester.assertRenderedPage(BudgetDetailsPage.class);
     }
-
-    private BudgetDetailData createBudget() {
-        BudgetDetailData budgetDetailData = new BudgetDetailData();
-        budgetDetailData.setId(1L);
-        budgetDetailData.setName("Budget 1");
-        budgetDetailData.setContractId(2L);
-        budgetDetailData.setContractName("Contract");
-        budgetDetailData.setTotal(Money.of(CurrencyUnit.EUR, 123));
-        budgetDetailData.setSpent(Money.of(CurrencyUnit.EUR, 43));
-        return budgetDetailData;
-    }
-
     @Override
     protected void setupTest() {
