+++ budgeteer-web-interface/src/test/java/org/wickedsource/budgeteer/web/pages/budgets/hours/BudgetHoursPageTest.java
@@ -11,19 +11,11 @@ import static org.mockito.Mockito.when;
 
 public class BudgetHoursPageTest extends AbstractWebTestTemplate {
 
-    @Autowired
-    private BudgetService service;
-
     @Test
     void render() {
         WicketTester tester = getTester();
-        when(service.loadBudgetBaseData(1L)).thenReturn(createBudget());
         tester.startPage(BudgetHoursPage.class, BudgetHoursPage.createParameters(1L));
         tester.assertRenderedPage(BudgetHoursPage.class);
-    }
-
-    private BudgetBaseData createBudget() {
-        return new BudgetBaseData(1L, "Budget 1");
     }
 
     @Override
