+++ budgeteer-web-interface/src/test/java/org/wickedsource/budgeteer/web/pages/person/weekreport/chart/TargetAndActualChartTest.java
@@ -13,36 +13,11 @@ import java.util.Random;
 
 public class TargetAndActualChartTest extends AbstractWebTestTemplate {
 
-    private Random random = new Random();
-
     @Test
     void testRender() {
         WicketTester tester = getTester();
         PersonWeeklyAggregationModel model = new PersonWeeklyAggregationModel(1L);
         tester.startComponentInPage(new TargetAndActualChart("chart", model, TargetAndActualChartConfiguration.Mode.WEEKLY));
-    }
-
-    public TargetAndActual createTargetAndActual(long personId, int numberOfWeeks) {
-        TargetAndActual targetAndActual = new TargetAndActual();
-
-        for (int i = 0; i < 5; i++) {
-            MoneySeries series = new MoneySeries();
-            series.setName("Budget " + i);
-            for (int j = 0; j < numberOfWeeks; j++) {
-                series.getValues().add(MoneyUtil.createMoneyFromCents(random.nextInt(5000)));
-            }
-            targetAndActual.getActualSeries().add(series);
-        }
-
-        MoneySeries series = new MoneySeries();
-        series.setName("Target");
-        for (int j = 0; j < numberOfWeeks; j++) {
-            series.getValues().add(MoneyUtil.createMoneyFromCents(random.nextInt(5000)));
-        }
-        targetAndActual.setTargetSeries(series);
-
-        return targetAndActual;
-
     }
 
     @Override
